refactor(room): move reaction code to utils
