+++ src/app/utils/matrix.ts
@@ -16,8 +16,8 @@ import {
 import to from 'await-to-js';
 import { IImageInfo, IThumbnailContent, IVideoInfo } from '$types/matrix/common';
 import { AccountDataEvent } from '$types/matrix/accountData';
-import { Membership, StateEvent } from '$types/matrix/room';
-import { getStateEvent } from './room';
+import { Membership, MessageEvent, StateEvent } from '$types/matrix/room';
+import { getEventReactions, getReactionContent, getStateEvent } from './room';
 
 const DOMAIN_REGEX = /\b(?:[a-zA-Z0-9-]+\.)+[a-zA-Z]{2,}\b/;
 
@@ -381,4 +381,30 @@ export const knockRestrictedSupported = (version: string): boolean => {
 export const creatorsSupported = (version: string): boolean => {
   const unsupportedVersion = ['1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11'];
   return !unsupportedVersion.includes(version);
+};
+
+export const toggleReaction = (
+  mx: MatrixClient,
+  room: Room,
+  targetEventId: string,
+  key: string,
+  shortcode?: string
+) => {
+  const relations = getEventReactions(room.getUnfilteredTimelineSet(), targetEventId);
+  const allReactions = relations?.getSortedAnnotationsByKey() ?? [];
+  const [, reactionsSet] = allReactions.find(([k]: [string, any]) => k === key) ?? [];
+  const reactions: MatrixEvent[] = reactionsSet ? Array.from(reactionsSet) : [];
+  const myReaction = reactions.find(factoryEventSentBy(mx.getUserId()!));
+
+  if (myReaction && !!(myReaction as any)?.isRelation()) {
+    mx.redactEvent(room.roomId, (myReaction as any).getId());
+    return;
+  }
+  const rShortcode =
+    shortcode || (reactions.find(eventWithShortcode)?.getContent().shortcode as string | undefined);
+  mx.sendEvent(
+    room.roomId,
+    MessageEvent.Reaction as any,
+    getReactionContent(targetEventId, key, rShortcode)
+  );
 };
